Refactor: ThreeMatchPuzzle cs
